feature(hash): refactor sha256 hash api (#953)
This is a WIP PR where I focus on improving performance around sha256
related things. The first commit re-organizes the API a bit to make it
easier to streamline the hashing later on. Single hashes do not need to
go through a incremental API and instead can be pipelined when done in a
loop, something to-be-implemented. And when doing mixins with just two
inputs of a known size, we can pipeline that as well, performing much
faster hashing than what we currently achieve.