fix(runtime): remove print from runtime test (#1002)
Please, no debug prints in tests.

It looks like this test was only added as a hack to satisfy codecov to
run lines of code that only exist for printing test outputs while
debugging tests locally. I don't think this is code we should cover in
our tests.

I can see how this helped get a prior PR merged, but now we should be
able to undo it.