feat(conformance): scripts to simplify running locally (#1001)
When I was first getting up and running with conformance, it was a
painful experience cloning all the right repos, checking out the right
versions, using the correct paths for all files, and running all the
correct commands. I spent several hours sorting out all these details.
The scripts in this PR are meant to save all these headaches so anyone
can bootstrap and run the entire conformance suite locally with minimal
effort.

see
[conformance/README.md](https://github.com/Syndica/sig/blob/dnut/conformance/run-scripts/conformance/README.md)

I didn't include any installation scripts for system dependencies
because these are very OS-specific, and it didn't give me much of a
problem when I was starting. On my machine running arch linux, I only
needed to install python3.11.

Also this fixes an issue in our CI. We weren't running the bpf loader v4
tests. Now we are.