fix(shred-network): flaky test (#1009)
For some reason this often takes over 2 seconds to complete in CI, which
causes it to fail because the test itself asserts that it must finish in
2 seconds. It is hard to comprehend why the test would ever take that
long. Running locally it finishes very fast.

I'd like a better solution for this than just increasing the timeout,
and I have a few ideas, but it's not a priority to implement right now.
We just need this to stop causing builds to fail. I lifted it from 2s to
10s which I'm hoping will significantly reduce the number of times this
fails in CI.


https://github.com/orgs/Syndica/projects/2/views/10?pane=issue&itemId=134153322