chore(testing): use random seed instead of hardcoding (#985)
It's preferred to use the random seed provided by the stdlib (which
changes every run, but is deterministic, and can be passed into the test
runner again to reproduce identical results), instead of just hardcoding
a single number. This ensures the unit doesn't have some hidden
dependence on the seed of the prng used.

Had to fix a couple cases just like this, where we would hardcode
expected results which depend on the seed to the prng. The change makes
it clearer what inputs are used to get said output.

Almost the entire diff is just that, maybe a few little fixes in random
places.