fix(grafana): log alert rule firing when there are no errors (#1059)
the alert should only fire within 5 minutes the last error, but this
alert remains in a perpetual firing status as soon as it sees one error.
the problem is that when the query evaluates to false (i.e. no error),
it actually becomes undefined, which grafana sees as "no data." no data
means no problem for us. noDataState is configured to KeepLast, which
means it persists the error status even when the error goes away.
changing to OK fixes this.

I also fixed the grafana url used in alert messages. it needs to include
port 3000