refactor(shred_network): organize into collector and transmitter (#1064)
These are two distinct pieces of the shred network that are only
connected with a channel and do not depend on any code from each other.
I find it useful to describe them with different language, and as
separate components, because usually only one of them is relevant at a
particular time.