perf(runtime): remove account cache & init SvmGateway per-slot (#1045)
I removed the BatchAccountCache struct and all account pre-loading
logic, and I replaced it with direct usages of the account store in the
runtime.

Now that there's no batch-specific state, I changed SvmGateway to only
load once per slot, instead of once per batch.

I consolidated CopiedAccount and CachedAccount into something called
LoadedAccount which represents any loaded accounts. For now, all
accounts are copied. This means frequently used accounts are copied
extra times. In the future, we should optimize this by eliminating
AccountSharedData and using Account throughout the runtime to avoid
copies by directly pointing to memory in the database, but this will be
a significant change.

I also simplified the return type from the transaction executor, which
enabled some cascading simplifications to other code in the committer
and the transaction harness. it returned a bunch of unnecessary
information that just made life harder for callers. this also revealed a
bug in solfuzz_agave, which i also needed to implement in our
transaction harness to ensure we have the same buggy behavior for
conformance.

I removed TransactionRollbacks and just replaced it with a list of
writes. Rollback logic has been inlined into check fee payer.

I ran a simple benchmark by running replay-offline on testnet for 60
seconds:
- master: 121 slots
- this branch: 125 slots