feat(identity): json private keys and base58 option for pubkeys (#1108)
1. Save private keys to disk as json instead of binary. This improves
compatibility with all the existing solana tooling that is built around
json key files and generally makes the data easier to work with since
it's plain text.
2. Allow configuration of vote account with either the private key file
path or the vote account address's base58 pubkey string. The private key
file is useful in the event that the operator still has the file after
generating it. But there's really no need for that file. You only need
the address. The operator may have deleted the private key because it's
no longer needed. In that case you can pass in the base58 encoded pubkey
on the cli and it still works. This is the same way agave works.