fix(gossip): Some refactors & fix silly LRU key collision (#1125)
Took the liberty of making a slew of small stylistic and code hygiene
changes, but the main important change here is in
`src/gossip/ping_pong.zig`, in the `maybePing` function.
Before, this function was using `std.testing.random_seed` to instantiate
a local `std.Random.DefaultPrng` for `Ping.initRandom`, which uses it to
generate a random value for its `token` field. This is problematic
because, while `std.testing.random_seed` will vary during tests, it will
always be set to zero outside of them, meaning that every invocation of
this function resulted in the generation of a `Ping` with the same exact
`token` every time, whose hash is used as the key for the entry in the
`pending_cache`, meaning that there would only ever be a single entry in
the `pending_cache` at any given time, leading to a false view of the
gossip table, where at most one contact info at a time would be up to
date.

The fix here is simply supplying a proper source of RNG, which is done
here by storing a PRNG named `ping_token_prng` in the gossip service
struct, and passing that down into `maybePing`, allowing
`filterValidPeers` to observe the true set of valid peers.