fix: reduce logging noise (#1132)
We keep seeing .TooFewTicks getting error logged in the alerts which is
unnecessary because it happens ordinarily on forks.

We also see error logs about a missing vote account when starting the
validator with consensus disabled, which makes no sense. Actually voting
should not even be enabled even when consensus is enabled, unless
explicitly enabled by the user. I turned it off by default, which
removes the error logs unless the user includes the option, at which
point the error would be justified.

To satisfy codecov I needed to add a way to test the logger so I added
`TestLogger` which is a logger implementation just for asserting what
gets logged.