fix(trace): better handle logger failing to allocate (#1117)
This is a quick fix for an issue in the logger where it fails to
allocate memory to send the message over the channel. Previously the
message would just get lost. I changed it to print it to stderr
directly, which solves the most pressing issue that messages are
disappearing.

We need some bigger changes to how the logger works here but it's not a
priority right now. The important thing is just making sure that the
message gets logged.

The codecov here is not being satisfied but I really don't see an
acceptable solution for this. It's not meaningful to test print
statements. It's also not acceptable to write to stderr directly in a
test because it clutters the output. I think we just need to bypass
codecov here.

Co-authored-by: Ahmad <12675427+ultd@users.noreply.github.com>