perf(thread): simplify `HomogeneousThreadPool` (#1070)
The HomogeneousThreadPool was inefficient with the use of
mutexes/conditions per task and spinning to join. It also leaked task
memory through `MemoryPool` if the caller never joins what was spawned.
This PR replaced it with a simpler API under the name
`ScopedThreadPool`, taking the function (instead of a task struct) for
convenience and solving the task leak with an internal free list.

There aren't many use cases of it. Eventually, we should remove
`ScopedThreadPool` and use `ThreadPool.schedule` directly as needed (an
example is the new `TransactionScheduler` merged in #1074 )