fix(shred-network): shred tracker assert fail index < bit_length
The problem is that the bit set has a capacity of `MAX_SHREDS_PER_SLOT/10` and we're seeing shreds with a higher index than this number. dividing by 10 was to avoid wasting memory and fragmenting data across memory. when i initially implemented this, mainnet slots never exceeded this value. apparently now we're seeing this number get exceeded, so we should probably just use the full theoretically max of MAX_SHREDS_PER_SLOT. this multiplies the memory usage by 10 so the struct is about 8 MB now. this is already being allocated.