refactor(gossip): improve startup time & misc (#1137)
Measured improvements:
* Increase the buffer size used in `buildPullRequests` which is passed
to `getThreadSafeGossipNodes`, aka `MAX_NUM_PULL_REQUESTS`. elements.
* Shorten `PULL_REQUEST_RATE` to 1 second, closer to agave's 500ms.

Other misc refactors:
* Store `entrypoints` as slice instead of arraylist.
* Remove some instances of `const Self = @This();`
* Make some parameters `*const`, and some unnecessary `var`s `const`.
* Add `Duration.mul`.
* Use some more decl literals.
* Unroll `VerifyMessageTask.run` and simplify it.
* Better static type safety for `getThreadSafeGossipNodes`.

---

Notes from some further investigation:

Originally the `PULL_REQUEST_RATE` was brought down to 500ms, to
directly match agave, but that seemingly lead seemingly somewhat more
volatile behavior, with the channels growing to rather huge sizes
(indicating a struggle to keep up with the increased push and pull of
gossip messages). An experiment in multi-threading the verifyMessage
tasks by @kprotty seemed to lighten the load quite a bit, so that's
probably worth looking into in the future.

Agave encodes imposes a limit on the number of push messages it will
send out ([`MAX_NUM_PUSHES = 1 <<
12`](https://github.com/anza-xyz/agave/blob/7fa607cf5a9a34b7ccbc9d881b4f0011d0edbdfe/gossip/src/crds_gossip_push.rs#L180)),
but does not appear to impose a limit on the number of pull requests it
sends out in the gossip logic itself, instead appearing to be bounded
only to
[1/16th](https://github.com/anza-xyz/agave/blob/7fa607cf5a9a34b7ccbc9d881b4f0011d0edbdfe/gossip/src/crds_gossip.rs#L342)
of the nodes it can send pull requests to, which is distinct from our
approach of having an actual strict cutoff, within which we send a
maximal number of pull requests.
May be worth taking inspiration from that approach, or experimenting
with other heuristics.